add change of decimation factor
